+++ RPG-Character-Sheets/src/br/com/while42/rpgcs/model/equip/weapons/AbstractWeapon.java
@@ -20,7 +20,7 @@ public abstract class AbstractWeapon implements Weapon {
 	}
 
 	@Override
-	public Integer getCost() {
+	public Long getCost() {
 		return base.getCost();
 	}
